First Functional Requirements.
This commit marks the creation of the initial set of functional requirements for the SigDespensa. These requirements outline the basic features and functionalities that the SigDespensa must possess in order to meet the needs of its users. The requirements were carefully considered and defined in order to ensure that they are clear, concise, and specific enough to guide the development of the SigDespensa. This commit serves as an important starting point for the development of the SigDespensa and provides a solid foundation for future work on this project.

Co-Authored-By: david santos <56547716+daviddevolin@users.noreply.github.com>
Co-Authored-By: Lucas Canto <91501668+lucascantodev@users.noreply.github.com>
Co-Authored-By: Isadora Stéfany <53875866+isadorazs@users.noreply.github.com>
Co-Authored-By: Manoel Anízio <92751858+manizio@users.noreply.github.com>